playlist/requirements.txt
aio-pika==9.5.5
redis==5.2.1
minio==7.2.15
pydantic==2.11.3
python-dotenv==1.1.0
aiohttp==3.10.11
